Move image fetching logic from page to CreationsGallery component
